web_map/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
requests==2.32.3
Pillow==10.4.0
python-dotenv==1.0.1
pixellab

